split macro crate into parser and macro
