+++ Cargo.toml
@@ -3,6 +3,7 @@ resolver = "2"
 members = [
     "leptos_i18n",
     "leptos_i18n_macro",
+    "leptos_i18n_parser",
     "tests/json",
     "tests/common",
     "tests/namespaces",
@@ -14,6 +15,7 @@ version = "0.5.0-gamma"
 
 [workspace.dependencies]
 leptos_i18n_macro = { path = "./leptos_i18n_macro", default-features = false, version = "0.5.0-gamma" }
+leptos_i18n_parser = { path = "./leptos_i18n_parser", default-features = false, version = "0.5.0-gamma" }
 leptos_i18n = { path = "./leptos_i18n", version = "0.5.0-gamma" }
 tests_common = { path = "./tests/common", version = "0.1.0" }
 leptos = { version = "0.7.0-gamma2" }
