+++ leptos_i18n_macro/Cargo.toml
@@ -24,14 +24,15 @@ toml = "0.8"
 icu = "1.5"
 fixed_decimal = { version = "0.5", features = ["ryu"] }
 json5 = { version = "0.4", optional = true }
+leptos_i18n_parser = { workspace = true, features = ["quote"] }
 
 [features]
 default = ["json_files"]
 nightly = []
 suppress_key_warnings = []
-json_files = []
-yaml_files = ["dep:serde_yaml"]
-json5_files = ["dep:json5"]
+json_files = ["leptos_i18n_parser/json_files"]
+yaml_files = ["dep:serde_yaml", "leptos_i18n_parser/yaml_files"]
+json5_files = ["dep:json5", "leptos_i18n_parser/json5_files"]
 interpolate_display = []
 track_locale_files = []
 experimental-islands = []
@@ -40,10 +41,10 @@ dynamic_load = []
 hydrate = []
 csr = []
 ssr = []
-plurals = []
-format_datetime = []
-format_list = []
-format_nums = []
+plurals = ["leptos_i18n_parser/plurals"]
+format_datetime = ["leptos_i18n_parser/format_datetime"]
+format_list = ["leptos_i18n_parser/format_list"]
+format_nums = ["leptos_i18n_parser/format_nums"]
 icu_compiled_data = []
 
 [package.metadata.cargo-all-features]
