+++ leptos_i18n_parser/.gitignore
@@ -1,0 +1,2 @@
+/target
+/Cargo.lock
