+++ leptos_i18n_parser/Cargo.toml
@@ -1,0 +1,32 @@
+[package]
+name = "leptos_i18n_parser"
+version = { workspace = true }
+edition = "2021"
+authors = ["Baptiste de Montangon"]
+license = "MIT"
+repository = "https://github.com/Baptistemontan/leptos_i18n"
+description = "parser for the leptos_i18n crate"
+readme = "../README.md"
+
+[dependencies]
+icu = "1.5"
+serde = { version = "1", features = ["rc"] }
+serde_json = { version = "1" }
+serde_yaml = { version = "0.9", optional = true }
+toml = "0.8"
+fixed_decimal = { version = "0.5", features = ["ryu"] }
+json5 = { version = "0.4", optional = true }
+quote = { version = "1", optional = true }
+syn = { version = "2.0", optional = true }
+proc-macro2 = { version = "1", optional = true }
+
+[features]
+default = ["json_files"]
+quote = ["dep:quote", "dep:syn", "dep:proc-macro2"]
+json_files = []
+yaml_files = ["dep:serde_yaml"]
+json5_files = ["dep:json5"]
+plurals = []
+format_datetime = []
+format_list = []
+format_nums = []
