+++ leptos_i18n_parser/src/lib.rs
@@ -1,0 +1,5 @@
+#![forbid(unsafe_code)]
+#![deny(warnings)]
+
+pub mod parse_locales;
+pub mod utils;
